refactor: split menu-screen file
